Left align comments
As a Data.govt.nz user I want comments to be presented in a clear and familiar way so I can read them and contribute easily.

Comments that appear on blog posts or data requests currently display in a ‘conversational’ format that goes back and forth between colours and right and left alignment. This can make it look like people are replying to each other when they aren’t, or as if two comments from one user are from multiple people.

Comments should always be left-aligned. Colour variation could be retained for clarity.

A.C.:
All comments wherever they appear on the site are left-aligned

ref - Jira ticket: https://govtnz.atlassian.net/browse/DATA-597